+++ scripts/docs-collator/ModuleRenderer.py
@@ -2,6 +2,7 @@ import logging
 import os
 import subprocess
 
+from Renderer import Renderer, TerraformDocsRenderingError
 from utils import io
 from utils import rendering, templating
 
@@ -23,13 +24,7 @@ INDEX_CATEGORY_TEMPLATE = jenv.get_template('index_category.json')
 SUBMODULE_TEMPLATE = jenv.get_template('submodule.readme.md')
 
 
-class TerraformDocsRenderingError(Exception):
-    def __init__(self, message):
-        self.message = message
-        super().__init__(f"Failed to render README.md. {message}")
-
-
-class ModuleRenderer:
+class ModuleRenderer(Renderer):
     def __init__(self, download_dir, docs_dir):
         self.download_dir = download_dir
         self.docs_dir = docs_dir
@@ -38,7 +33,7 @@ class ModuleRenderer:
         logging.info(f"Rendering doc for: {repo.full_name}")
         module_download_dir = os.path.join(self.download_dir, repo.name)
 
-        self.__pre_rendering_fixes(repo, module_download_dir)
+        self._pre_rendering_fixes(repo, module_download_dir)
 
         provider, module_name = rendering.parse_terraform_repo_name(repo.name)
         module_docs_dir = os.path.join(self.docs_dir, provider, module_name)
@@ -49,9 +44,9 @@ class ModuleRenderer:
         io.copy_file(readme_md_file, os.path.join(module_docs_dir, README_MD))
 
         readme_md_file = os.path.join(module_docs_dir, README_MD)
-        self.__post_rendering_fixes(repo, readme_md_file)
+        self._post_rendering_fixes(repo, readme_md_file)
 
-        self.__copy_extra_resources_for_docs(module_download_dir, module_docs_dir)
+        self._copy_extra_resources_for_docs(module_download_dir, module_docs_dir)
         self.__copy_extra_resources_for_images(module_download_dir, module_docs_dir)
         self.__copy_extra_resources_for_submodules(repo, module_download_dir, module_docs_dir)
 
@@ -77,19 +72,6 @@ class ModuleRenderer:
             raise TerraformDocsRenderingError(error_message)
 
         logging.info(f"Rendered: {readme_md_file}")
-
-    def __copy_extra_resources_for_docs(self, module_download_dir, module_docs_dir):
-        extra_resources_dir = os.path.join(module_download_dir, DOCS_DIR)
-        files = io.get_filenames_in_dir(extra_resources_dir, '*', True)
-
-        for file in files:
-            if os.path.basename(file).lower().endswith('.md') or os.path.isdir(file):
-                continue
-
-            dest_file = os.path.join(module_docs_dir, DOCS_DIR, os.path.relpath(file, extra_resources_dir))
-            io.copy_file(file, dest_file)
-
-            logging.info(f"Copied extra file: {dest_file}")
 
     def __copy_extra_resources_for_images(self, module_download_dir, module_docs_dir):
         extra_resources_dir = os.path.join(module_download_dir, IMAGES_DIR)
@@ -154,21 +136,6 @@ class ModuleRenderer:
                                                  title=name)
 
         io.save_string_to_file(os.path.join(dir, INDEX_CATEGORY_JSON), content)
-
-    def __pre_rendering_fixes(self, repo, module_download_dir):
-        readme_yaml_file = os.path.join(module_download_dir, README_YAML)
-        content = io.read_file_to_string(readme_yaml_file)
-        content = rendering.remove_targets_md(content)
-        content = rendering.rename_name(repo, content)
-        io.save_string_to_file(readme_yaml_file, content)
-
-    def __post_rendering_fixes(self, repo, readme_md_file):
-        content = io.read_file_to_string(readme_md_file)
-        content = rendering.fix_self_non_closing_br_tags(content)
-        content = rendering.fix_custom_non_self_closing_tags_in_pre(content)
-        content = rendering.fix_github_edit_url(content, repo)
-        content = rendering.fix_sidebar_label(content, repo)
-        io.save_string_to_file(readme_md_file, content)
 
     def __post_rendering_fixes_for_submodule(self, readme_md_file):
         content = io.read_file_to_string(readme_md_file)
